qshield-backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
sqlalchemy==2.0.48
pydantic==2.12.5
email-validator==2.2.0
passlib==1.7.4
python-multipart==0.0.22
starlette==1.0.0
python-jose[cryptography]
cryptography
PyJWT==2.12.1
bcrypt==5.0.0
python-dotenv
APScheduler==3.10.4
pyotp==2.9.0
qrcode==8.2
